+++ app/src/main/kotlin/info/dvkr/switchmovie/viewmodel/moviegrid/MovieGridViewEvent.kt
@@ -6,8 +6,8 @@ sealed class MovieGridViewEvent : BaseViewModel.Event {
     object Refresh : MovieGridViewEvent()
     object LoadMore : MovieGridViewEvent()
     object Update : MovieGridViewEvent()
-    data class SetMovieStar(val movieId: Int) : MovieGridViewEvent()
-    data class UnsetMovieStar(val movieId: Int) : MovieGridViewEvent()
+    data class SetMovieStar(val movieId: Long) : MovieGridViewEvent()
+    data class UnsetMovieStar(val movieId: Long) : MovieGridViewEvent()
 
     override fun toString(): String = this::class.java.simpleName
 }
