+++ domain/src/main/kotlin/info/dvkr/switchmovie/domain/repositories/MovieRepository.kt
@@ -1,10 +1,9 @@
 package info.dvkr.switchmovie.domain.repositories
 
-import androidx.lifecycle.LiveData
 import info.dvkr.switchmovie.domain.model.Movie
 import info.dvkr.switchmovie.domain.utils.Either
 import kotlinx.coroutines.flow.Flow
-import org.threeten.bp.LocalDate
+import java.time.LocalDate
 
 
 interface MovieRepository {
@@ -12,9 +11,9 @@ interface MovieRepository {
 
         fun getMovies(): Flow<List<Movie>>
 
-        suspend fun getMovieById(movieId: Int): Movie?
+        suspend fun getMovieById(movieId: Long): Movie?
 
-        fun getMovieFlowById(movieId: Int): Flow<Movie>
+        fun getMovieFlowById(movieId: Long): Flow<Movie>
 
         fun getLastMovieUpdateDate(): LocalDate
     }
